% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pzfx.R
\name{as_column_table}
\alias{as_column_table}
\title{Convert the tibble to a pzfx table XML node}
\usage{
as_column_table(data, row_names = NULL, table_name = NULL)
}
\arguments{
\item{data}{tibble to be converted}

\item{row_names}{if `NULL` (the default), converts the data frame row names to Prism row titles. If `character`, uses the provided column to generate the row titles.}

\item{table_name}{`character` to specify the Prism table title. if `NULL` (the default), will use the name of `data`.}
}
\value{
A xml2 xml document
}
\description{
Convert a **wide** tibble to a pzfx table XML node.
}
