% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkside.R
\name{pzfx_tables}
\alias{pzfx_tables}
\title{List all data tables in a pzfx file}
\usage{
pzfx_tables(path)
}
\arguments{
\item{path}{Path to the pzfx file.}
}
\value{
A character vector with the name of the data tables
}
\description{
List all data tables in a pzfx file
}
