% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkside.R
\name{read_pzfx}
\alias{read_pzfx}
\title{Read Graphpad prism pzfx files}
\usage{
read_pzfx(path, data_table = 1)
}
\arguments{
\item{path}{Path to the pzfx file.}

\item{data_table}{data table to read. Either a string (the name of a data table), or an integer (the position of the data table). Defaults to the first data table.}
}
\value{
A tibble / data frame
}
\description{
`read_pzfx()` generates a tidy data frame out of the data table contained in the pzfx file.
}
