% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pzfx.R
\name{write_pzfx}
\alias{write_pzfx}
\title{Write a data frame to a Graphpad prism pzfx file}
\usage{
write_pzfx(data, path, table_type = "column", table_name = NULL,
  row_names = NULL)
}
\arguments{
\item{data}{tibble to be converted}

\item{path}{to write to}

\item{table_type}{Prism table type (column)}

\item{table_name}{Name of the Prism table in the output document. if `NULL` (the default) the name of `data` is used.}

\item{row_names}{Name of column to use for rownames. If `NULL` (the default), uses row names if they exist.}
}
\value{
A xml2 xml document
}
\description{
Writes out a `pzfx` file which can be opened using Graphpad Prism.
}
